+++ out/extension.js
@@ -6,27 +6,60 @@ const parser_1 = require("./parser");
 function activate(context) {
     let activeEditor;
     let parser = new parser_1.Parser();
-    let updateDecorations = function (useHash = false) {
+    let currentLanguageId;
+    context.subscriptions.push(parser);
+    let updateDecorations = function () {
         if (!activeEditor)
             return;
-        if (!parser.supportedLanguage)
-            return;
+        setParserLanguage(activeEditor);
         parser.ClearDecorations();
+        if (!parser.supportedLanguage) {
+            parser.ApplyDecorations(activeEditor);
+            parser.UpdateCursorDecorations(activeEditor);
+            return;
+        }
         parser.FindSingleLineComments(activeEditor);
         parser.FindBlockComments(activeEditor);
         parser.FindJSDocComments(activeEditor);
         parser.ApplyDecorations(activeEditor);
         parser.UpdateCursorDecorations(activeEditor);
     };
+    function setParserLanguage(editor) {
+        const languageChanged = currentLanguageId !== editor.document.languageId;
+        currentLanguageId = editor.document.languageId;
+        parser.SetRegex(currentLanguageId);
+        return languageChanged;
+    }
+    function clearEditorDecorations(editor) {
+        parser.ClearDecorations();
+        parser.ApplyDecorations(editor);
+        parser.UpdateCursorDecorations(editor);
+    }
     if (vscode.window.activeTextEditor) {
         activeEditor = vscode.window.activeTextEditor;
-        parser.SetRegex(activeEditor.document.languageId);
+        setParserLanguage(activeEditor);
+        clearEditorDecorations(activeEditor);
         triggerUpdateDecorations();
     }
     vscode.window.onDidChangeActiveTextEditor((editor) => {
         activeEditor = editor;
         if (editor) {
-            parser.SetRegex(editor.document.languageId);
+            setParserLanguage(editor);
+            clearEditorDecorations(editor);
+            triggerUpdateDecorations();
+        }
+        else {
+            currentLanguageId = undefined;
+            parser.ClearDecorations();
+        }
+    }, null, context.subscriptions);
+    vscode.workspace.onDidOpenTextDocument(document => {
+        if (activeEditor && document.uri.toString() === activeEditor.document.uri.toString()) {
+            triggerUpdateDecorations();
+        }
+    }, null, context.subscriptions);
+    vscode.workspace.onDidCloseTextDocument(document => {
+        if (activeEditor && document.uri.toString() === activeEditor.document.uri.toString()) {
             triggerUpdateDecorations();
         }
     }, null, context.subscriptions);
@@ -37,6 +70,9 @@ function activate(context) {
     }, null, context.subscriptions);
     vscode.window.onDidChangeTextEditorSelection(event => {
         if (activeEditor && event.textEditor === activeEditor) {
+            if (setParserLanguage(activeEditor)) {
+                clearEditorDecorations(activeEditor);
+            }
             parser.UpdateCursorDecorations(activeEditor);
         }
     }, null, context.subscriptions);